Split renderer extensions into its own module
