+++ src/lib.rs
@@ -27,7 +27,8 @@ extern crate draw_state;
 /// public re-exported traits
 pub mod traits {
     pub use device::{Device, Factory};
-    pub use render::device_ext::{DeviceExt, FactoryExt};
+    pub use render::ext::device::DeviceExt;
+    pub use render::ext::factory::FactoryExt;
     pub use render::mesh::ToSlice;
 }
 
@@ -37,7 +38,8 @@ pub use draw_state::{DrawState, BlendPreset};
 // public re-exports
 pub use render::{Renderer, DrawError};
 pub use render::batch;
-pub use render::device_ext::{Graphics, ShaderSource, ProgramError};
+pub use render::ext::device::Graphics;
+pub use render::ext::shade::{ShaderSource, ProgramError};
 pub use render::mesh::{Attribute, Mesh, VertexFormat};
 pub use render::mesh::Error as MeshError;
 pub use render::mesh::{Slice, ToSlice, SliceKind};
